requirements.txt
Django==4.0.1
psycopg2-binary
djangorestframework==3.14.0
pydantic==1.10.8
requests==2.30.0
django-environ==0.10.0